Fixes heel effect when successfully removing devious boots
Currently when the paired animation to remove devious boots ends up successfully remove them, the player may "sunk into the ground". This is due to `_RestoreHeelEffect` returning directly when there is no boots equipped, causing the negative override added during the animation not properly removed.

This fix removes the negative override first, regardless if there is still an equipped boot or not.